+++ analysis/src/BuildSystem.ml
@@ -6,15 +6,18 @@ let namespacedName namespace name =
 let ( /+ ) = Filename.concat
 
 let getBsPlatformDir rootPath =
-  let result =
-    ModuleResolution.resolveNodeModulePath ~startPath:rootPath "rescript"
-  in
-  match result with
-  | Some path -> Some path
-  | None ->
-    let message = "rescript could not be found" in
-    Log.log message;
-    None
+  match !Cfg.isDocGenFromCompiler with
+  | false -> (
+    let result =
+      ModuleResolution.resolveNodeModulePath ~startPath:rootPath "rescript"
+    in
+    match result with
+    | Some path -> Some path
+    | None ->
+      let message = "rescript could not be found" in
+      Log.log message;
+      None)
+  | true -> Some rootPath
 
 let getLibBs root = Files.ifExists (root /+ "lib" /+ "bs")