+++ analysis/src/SharedTypes.ml
@@ -311,7 +311,11 @@ end = struct
     {env with exported = structure.exported; pathRev; parent = Some env}
 end
 
-type polyVariantConstructor = {name: string; args: Types.type_expr list}
+type polyVariantConstructor = {
+  name: string;
+  displayName: string;
+  args: Types.type_expr list;
+}
 
 type innerType = TypeExpr of Types.type_expr | ExtractedType of completionType
 and completionType =
@@ -493,6 +497,7 @@ type package = {
   builtInCompletionModules: builtInCompletionModules;
   opens: path list;
   uncurried: bool;
+  rescriptVersion: int * int;
 }
 
 let allFilesInPackage package =
@@ -732,6 +737,16 @@ module Completable = struct
     | ChtmlElement {prefix} -> "ChtmlElement <" ^ prefix
 end
 
+module ScopeTypes = struct
+  type item =
+    | Constructor of string * Location.t
+    | Field of string * Location.t
+    | Module of string * Location.t
+    | Open of string list
+    | Type of string * Location.t
+    | Value of string * Location.t * Completable.contextPath option * item list
+end
+
 module Completion = struct
   type kind =
     | Module of Module.t
@@ -745,7 +760,7 @@ module Completion = struct
     | FileModule of string
     | Snippet of string
     | ExtractedType of completionType * [`Value | `Type]
-    | FollowContextPath of Completable.contextPath
+    | FollowContextPath of Completable.contextPath * ScopeTypes.item list
 
   type t = {
     name: string;